+++ src/configs/esm.js
@@ -4,13 +4,7 @@ module.exports = {
 	parserOptions: {
 		requireConfigFile: false,
 		sourceType: "module",
-		babelOptions: {
-			plugins: [
-				require("fs").existsSync("node_modules/@babel/plugin-syntax-top-level-await")
-					? "@babel/plugin-syntax-top-level-await"
-					: "@onedotprojects/eslint-plugin/node_modules/@babel/plugin-syntax-top-level-await",
-			],
-		},
+		babelOptions: {},
 	},
 	rules: {
 		"id-match": [